requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
japanize-matplotlib==1.1.3
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.14.1
six==1.16.0
